requirements.txt
Django==2.1.7
django-tinymce4-lite==1.7.4
jsmin==2.2.2
Pillow==5.4.1
pytz==2018.9
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-tinymce4-lite==1.7.4
idna==2.8
jsmin==2.2.2
oauthlib==3.0.1
Pillow==5.4.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
